Massive Gui Reworks and Machine Process Changes.
The gold tanks inside of the gold forge and the coin press now require
constant power per tick to stay liquid, if this power requirement is not
met for more then a minute, the machines will enter the "Frozen" state.
a "Frozen" machine will not forge any new gold/press coins again until
it has gone thru the "Warming" state. Time spent warming grows based on
the amount of gold stored in the tank.